Add PDF generation script and refactor resume rendering
