+++ src/Columns/Highlight.jsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { Text, View } from "@react-pdf/renderer";
 import { style } from "../Styles";
 import { Section } from "./Components/Section";
